Move AST into separate directory. Small cleanups.
