+++ src/compiler/lexer.cpp
@@ -8,9 +8,7 @@
  * \license Distributed under the MIT software license (see accompanying LICENSE.txt).
  */
 
-#include <algorithm>
 #include <array>
-#include <stdexcept>
 
 #include <fmt/core.h>
