+++ test/CMakeLists.txt
@@ -39,8 +39,9 @@ add_test(NAME lexer COMMAND test_lexer)
 add_executable(test_parser
     test_parser.cpp
     ../src/compiler/builtins/format.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/lexer.cpp
     ../src/compiler/parser.cpp
@@ -58,8 +59,9 @@ add_test(NAME parser COMMAND test_parser)
 add_executable(test_codegen
     test_codegen.cpp
     ../src/compiler/builtins/format.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/lexer.cpp
     ../src/compiler/parser.cpp
@@ -77,8 +79,9 @@ add_test(NAME codegen COMMAND test_codegen)
 add_executable(test_compile_ir
     test_compile_ir.cpp
     ../src/compiler/builtins/format.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/lexer.cpp
     ../src/compiler/parser.cpp
@@ -97,8 +100,9 @@ add_executable(test_type_system
     test_type_system.cpp
     ../src/compiler/builtins/format.cpp
     ../src/archives/file.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/lexer.cpp
     ../src/compiler/parser.cpp
@@ -124,8 +128,9 @@ add_executable(test_resolve
     test_resolve.cpp
     ../src/compiler/builtins/format.cpp
     ../src/archives/file.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/lexer.cpp
     ../src/compiler/parser.cpp
@@ -152,8 +157,9 @@ add_executable(test_output
     test_output.cpp
     ../src/compiler/builtins/format.cpp
     ../src/archives/file.cpp
-    ../src/compiler/ast.cpp
-    ../src/compiler/eval.cpp
+    ../src/compiler/ast/ast.cpp
+    ../src/compiler/ast/eval.cpp
+    ../src/compiler/ast/macros.cpp
     ../src/compiler/codegen.cpp
     ../src/compiler/emitter.cpp
     ../src/compiler/lexer.cpp
